Tests for services passing

+++ src/Azure.IIoT.OpcUa.Testing/src/Fixtures/BaseServerFixture.cs
@@ -8,16 +8,20 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
     using Azure.IIoT.OpcUa.Publisher.Stack;
     using Azure.IIoT.OpcUa.Publisher.Stack.Sample;
     using Azure.IIoT.OpcUa.Publisher.Stack.Services;
+    using Azure.IIoT.OpcUa.Shared.Models;
     using Azure.IIoT.OpcUa.Testing.Runtime;
     using Furly.Extensions.Logging;
     using Furly.Extensions.Serializers.Json;
-    using Furly.Extensions.Serializers.Newtonsoft;
     using Furly.Extensions.Utils;
     using Microsoft.Extensions.Logging;
+    using Opc.Ua;
     using Opc.Ua.Server;
     using System;
     using System.Collections.Generic;
     using System.IO;
+    using System.Linq;
+    using System.Net;
+    using System.Net.Sockets;
     using System.Security.Cryptography.X509Certificates;
     using System.Threading;
     using System.Threading.Tasks;
@@ -27,6 +31,11 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
     /// </summary>
     public abstract class BaseServerFixture : IDisposable
     {
+        /// <summary>
+        /// Host server is running on
+        /// </summary>
+        public IPHostEntry Host { get; }
+
         /// <summary>
         /// Port server is listening on
         /// </summary>
@@ -40,7 +49,7 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
         /// <summary>
         /// Cert folder
         /// </summary>
-        public string PkiRootPath { get; private set; }
+        public string PkiRootPath { get; }
 
         /// <summary>
         /// Logger
@@ -53,6 +62,34 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
         public OpcUaClientManager Client => _client.Value;
 
         /// <summary>
+        /// Start port
+        /// </summary>
+        public static void SetStartPort(int value)
+        {
+            _nextPort = value;
+        }
+
+        /// <summary>
+        /// Get server connection
+        /// </summary>
+        /// <returns></returns>
+        public ConnectionModel GetConnection(string path = null)
+        {
+            return new ConnectionModel
+            {
+                Endpoint = new EndpointModel
+                {
+                    Url = $"opc.tcp://{Host?.HostName ?? "localhost"}:{Port}/{path ?? "UA/SampleServer"}",
+                    AlternativeUrls = Host?.AddressList
+                        .Where(ip => ip.AddressFamily == AddressFamily.InterNetwork)
+                        .Select(ip => $"opc.tcp://{ip}:{Port}/{path ?? "UA/SampleServer"}")
+                        .ToHashSet(),
+                    Certificate = Certificate?.RawData?.ToThumbprint()
+                }
+            };
+        }
+
+        /// <summary>
         /// Create fixture
         /// </summary>
         protected BaseServerFixture(IEnumerable<INodeManagerFactory> nodes)
@@ -61,6 +98,8 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
             {
                 throw new ArgumentNullException(nameof(nodes));
             }
+            Host = Try.Op(() => Dns.GetHostEntry(Utils.GetHostName()))
+                ?? Try.Op(() => Dns.GetHostEntry("localhost"));
             Logger = Log.Console<BaseServerFixture>(LogLevel.Debug);
             _config = new TestClientServicesConfig();
             _client = new Lazy<OpcUaClientManager>(() => new OpcUaClientManager(Logger, _config, new DefaultJsonSerializer()), false);
@@ -98,26 +137,50 @@ namespace Azure.IIoT.OpcUa.Testing.Fixtures
         /// <inheritdoc/>
         public void Dispose()
         {
-            Logger.LogInformation("Disposing server and client fixture...");
-            _serverHost.Dispose();
-            // Clean up all created certificates
-            if (Directory.Exists(PkiRootPath))
+            // Do not change this code. Put cleanup code in 'Dispose(bool disposing)' method
+            Dispose(disposing: true);
+            GC.SuppressFinalize(this);
+        }
+
+        /// <summary>
+        /// Override to dispose
+        /// </summary>
+        /// <param name="disposing"></param>
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!_disposedValue)
             {
-                Logger.LogInformation("Server disposed - cleaning up server certificates...");
-                Try.Op(() => Directory.Delete(PkiRootPath, true));
+                if (disposing)
+                {
+                    Logger.LogInformation("Disposing server and client fixture...");
+                    _serverHost.Dispose();
+                    // Clean up all created certificates
+                    if (Directory.Exists(PkiRootPath))
+                    {
+                        Logger.LogInformation("Server disposed - cleaning up server certificates...");
+                        Try.Op(() => Directory.Delete(PkiRootPath, true));
+                    }
+                    if (_client.IsValueCreated)
+                    {
+                        Logger.LogInformation("Disposing client...");
+                        Task.Run(() => _client.Value.Dispose()).Wait();
+                    }
+                    Logger.LogInformation("Client disposed - cleaning up client certificates...");
+                    _config?.Dispose();
+                    Logger.LogInformation("Server and client fixture disposed.");
+                }
+
+                // TODO: free unmanaged resources (unmanaged objects) and override finalizer
+                // TODO: set large fields to null
+                _disposedValue = true;
             }
-            if (_client.IsValueCreated)
-            {
-                Logger.LogInformation("Disposing client...");
-                Task.Run(() => _client.Value.Dispose()).Wait();
-            }
-            Logger.LogInformation("Client disposed - cleaning up client certificates...");
-            _config?.Dispose();
-            Logger.LogInformation("Server and client fixture disposed.");
         }
 
         private static readonly Random kRand = new();
+#pragma warning disable CA5394 // Do not use insecure randomness
         private static volatile int _nextPort = kRand.Next(53000, 58000);
+#pragma warning restore CA5394 // Do not use insecure randomness
+        private bool _disposedValue;
         private readonly IServerHost _serverHost;
         private readonly TestClientServicesConfig _config;
         private readonly Lazy<OpcUaClientManager> _client;
